octofit-tracker/backend/requirements.txt
# Updated dependencies to resolve compatibility issues
Django>=4.2
djangorestframework>=3.16.0
django-allauth==0.51.0
django-cors-headers>=4.7.0
dj-rest-auth
# Removed djongo and added mongoengine for MongoDB integration
mongoengine>=0.27.0
pymongo==3.12
# Updated sqlparse to a compatible version
sqlparse>=0.3.1
stack-data==0.6.3
sympy==1.12
tenacity==9.0.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.3.0
tornado==6.4.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20240906
typing_extensions==4.9.0
tzdata==2024.2
uri-template==1.3.0
urllib3==2.2.3
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
